astro_project/requirements.txt
apache-airflow-providers-postgres
pandas
requests
soda-core-postgres
setuptools
sqlalchemy
psycopg2-binary